# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Cooking Companion API
  version: 0.1.0
  description: >
    MVP surface from MON-12. Mobile generates a typed client from this spec
    (`openapi-typescript` or `openapi-generator-cli`).
    Import preview/commit bodies and JSON match Spring `ImportController` DTOs
    (`RecipeDraftResponse`, `ImportCommitRequest`, `VariantDetailResponse`).
    UI-oriented nested `draft` shapes in the app are normalized in `src/api/client.ts`.
    **Households:** `GET /api/v1/households` and `POST /api/v1/households/join` require an authenticated
    principal (`Authorization: Bearer` JWT with `sub` UUID, or dev-only `X-User-Id`). Recipe reads/writes
    for a shared library use optional `X-Household-Id` (must match membership).
servers:
  - url: https://api.example.com
    description: Production
paths:
  /api/v1/dishes:
    get:
      operationId: listDishes
      tags: [Dishes]
      summary: List dishes
      parameters:
        - name: q
          in: query
          required: false
          schema:
            type: string
          description: >
            Optional search on dish title. Case-insensitive substring match.
            Absent or whitespace-only values are ignored (full list).
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Dish'
    post:
      operationId: createDish
      tags: [Dishes]
      summary: Create dish
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DishCreate'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dish'
  /api/v1/dishes/{dishId}:
    get:
      operationId: getDish
      tags: [Dishes]
      summary: Get dish by id
      parameters:
        - name: dishId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dish'
    patch:
      operationId: patchDish
      tags: [Dishes]
      summary: Update dish
      parameters:
        - name: dishId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchDish'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dish'
    delete:
      operationId: deleteDish
      tags: [Dishes]
      summary: Delete dish (cascades variants)
      parameters:
        - name: dishId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: No Content
  /api/v1/dishes/{dishId}/variants:
    get:
      operationId: listVariants
      tags: [Dish variants]
      summary: List variants for dish
      parameters:
        - name: dishId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RecipeVariantSummary'
    post:
      operationId: createVariant
      tags: [Dish variants]
      summary: Create variant
      parameters:
        - name: dishId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecipeVariantCreate'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecipeVariantSummary'
  /api/v1/variants/{variantId}:
    get:
      operationId: getVariant
      tags: [Variants]
      summary: Get variant with ingredients and steps
      parameters:
        - name: variantId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VariantDetailResponse'
    patch:
      operationId: patchVariant
      tags: [Variants]
      summary: Patch variant metadata and optionally replace ingredients/steps
      parameters:
        - name: variantId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchVariant'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VariantDetailResponse'
    delete:
      operationId: deleteVariant
      tags: [Variants]
      summary: Delete variant
      parameters:
        - name: variantId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: No Content
  /api/v1/variants/{variantId}/fork:
    post:
      operationId: forkVariant
      tags: [Variants]
      summary: Fork variant (copy-on-write under same dish)
      parameters:
        - name: variantId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VariantDetailResponse'
  /api/v1/recipe-ai/flags:
    get:
      operationId: getRecipeAiFlags
      summary: Recipe AI feature flags
      tags: [Recipe AI]
      description: Clients use this to gate optional generative adjustment UX before calling apply-profile.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecipeAiFlags'
  /api/v1/variants/{variantId}/apply-profile:
    post:
      operationId: applyVariantProfile
      summary: Apply parameter profile
      tags: [Variants]
      description: >
        Preview-only: returns adjusted ingredients/steps and persists a VariantAdjustment audit row; does not overwrite the variant.
        Deterministic rules: `dairyMode` = none | omit | substitute_oat; optional `omitTokens`.
        Optional `useGenerative: true` (explicit opt-in) requires `GET /api/v1/recipe-ai/flags` and provider config; otherwise 403/503.
      parameters:
        - name: variantId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApplyProfileRequest'
      responses:
        '200':
          description: Preview result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplyProfileResponse'
  /api/v1/households:
    get:
      operationId: listHouseholds
      summary: List households the authenticated user belongs to
      tags: [Households]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/HouseholdSummary'
  /api/v1/households/join:
    post:
      operationId: joinHousehold
      summary: Join a household with an invite code
      tags: [Households]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HouseholdJoinRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HouseholdSummary'
  /api/v1/import/preview:
    post:
      operationId: importPreview
      tags: [Import]
      summary: Preview import from URL or HTML (no persist)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImportPreviewRequest'
      responses:
        '200':
          description: OK — wire shape `RecipeDraftResponse` (flat; not nested under `draft`)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecipeDraftResponse'
  /api/v1/import/commit:
    post:
      operationId: importCommit
      tags: [Import]
      summary: Commit validated draft
      parameters:
        - name: Idempotency-Key
          in: header
          required: false
          schema:
            type: string
          description: >
            When set, identical replays return the same persisted variant (see integration tests).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImportCommitRequest'
      responses:
        '201':
          description: Created — full variant detail (`VariantDetailResponse`), same as GET /variants/{id}
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VariantDetailResponse'
components:
  schemas:
    HouseholdSummary:
      type: object
      required: [id, name]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        membershipRole:
          type: string
          nullable: true
          description: owner, member, …
    HouseholdJoinRequest:
      type: object
      required: [code]
      properties:
        code:
          type: string
    Dish:
      type: object
      required: [id, name]
      properties:
        id:
          type: string
        name:
          type: string
        tags:
          type: array
          items:
            type: string
        heroImageUrl:
          type: string
          nullable: true
    DishCreate:
      type: object
      required: [name]
      properties:
        name:
          type: string
        tags:
          type: array
          items:
            type: string
    PatchDish:
      type: object
      properties:
        name:
          type: string
        tags:
          type: array
          items:
            type: string
        heroImageUrl:
          type: string
          nullable: true
    PatchVariant:
      type: object
      properties:
        title:
          type: string
        yields:
          type: string
        prepTimeMin:
          type: integer
        cookTimeMin:
          type: integer
        canonical:
          type: boolean
        sourceId:
          type: string
          format: uuid
        ingredients:
          type: array
          items:
            $ref: '#/components/schemas/IngredientLineDto'
        steps:
          type: array
          items:
            $ref: '#/components/schemas/RecipeStepDto'
    RecipeVariantSummary:
      type: object
      required: [id, dishId, title]
      properties:
        id:
          type: string
        dishId:
          type: string
        title:
          type: string
        yields:
          type: string
        totalTimeMin:
          type: integer
          nullable: true
        isCanonical:
          type: boolean
        sourceAttribution:
          type: string
          nullable: true
    IngredientLine:
      type: object
      required: [id, text]
      properties:
        id:
          type: string
        text:
          type: string
        quantity:
          type: object
          nullable: true
          properties:
            amount:
              type: number
            unit:
              type: string
    RecipeStep:
      type: object
      required: [id, order, text]
      properties:
        id:
          type: string
        order:
          type: integer
        text:
          type: string
        timerSec:
          type: integer
          nullable: true
    Source:
      type: object
      required: [type]
      properties:
        type:
          type: string
          enum: [web, manual, import_file, youtube]
        url:
          type: string
          nullable: true
        attribution:
          type: string
    RecipeVariantCreate:
      type: object
      required: [title, ingredients, steps]
      properties:
        title:
          type: string
        yields:
          type: string
        totalTimeMin:
          type: integer
          nullable: true
        ingredients:
          type: array
          items:
            $ref: '#/components/schemas/IngredientLine'
        steps:
          type: array
          items:
            $ref: '#/components/schemas/RecipeStep'
        source:
          $ref: '#/components/schemas/Source'
          nullable: true
    ImportPreviewRequest:
      type: object
      properties:
        url:
          type: string
        html:
          type: string
    IngredientLineDto:
      type: object
      required: [sortOrder, ingredientText]
      properties:
        id:
          type: string
          nullable: true
          description: Stable id when reading a persisted variant; omit on create
        sortOrder:
          type: integer
        amountNumeric:
          type: number
          nullable: true
        unit:
          type: string
          nullable: true
        ingredientText:
          type: string
        preparationNote:
          type: string
          nullable: true
        alternates:
          type: array
          items:
            type: string
    RecipeStepDto:
      type: object
      required: [sortOrder, text]
      properties:
        id:
          type: string
          nullable: true
        sortOrder:
          type: integer
        text:
          type: string
        timerSeconds:
          type: integer
          nullable: true
        linkUrl:
          type: string
          nullable: true
    CreateVariantRequest:
      type: object
      required: [title, canonical]
      properties:
        title:
          type: string
        yields:
          type: string
          nullable: true
        prepTimeMin:
          type: integer
          nullable: true
        cookTimeMin:
          type: integer
          nullable: true
        canonical:
          type: boolean
        sourceId:
          type: string
          format: uuid
          nullable: true
        ownerUserId:
          type: string
          format: uuid
          nullable: true
        ingredients:
          type: array
          items:
            $ref: '#/components/schemas/IngredientLineDto'
        steps:
          type: array
          items:
            $ref: '#/components/schemas/RecipeStepDto'
    RecipeDraftResponse:
      type: object
      required: [confidence, warnings, parseMethod, variantDraft]
      properties:
        suggestedDishName:
          type: string
          nullable: true
        confidence:
          type: number
          description: Parse confidence 0..1
        warnings:
          type: array
          items:
            type: string
        heroImageUrl:
          type: string
          nullable: true
        parseMethod:
          type: string
          description: json_ld, microdata, or heuristic
        previewId:
          type: string
          format: uuid
          nullable: true
          description: Set after server persists preview row; send on commit to merge draft
        variantDraft:
          $ref: '#/components/schemas/CreateVariantRequest'
    ImportCommitRequest:
      type: object
      description: >
        When `previewId` is set, server loads stored preview; `dishName` and `variant` default from it.
        Otherwise provide `dishName` and `variant` (validated).
      properties:
        previewId:
          type: string
          format: uuid
          nullable: true
        dishName:
          type: string
          nullable: true
        dishTags:
          type: array
          items:
            type: string
        heroImageUrl:
          type: string
          nullable: true
        ownerUserId:
          type: string
          format: uuid
          nullable: true
        sourceUrl:
          type: string
          nullable: true
          description: Source URL for idempotency when ownerUserId is set
        variant:
          $ref: '#/components/schemas/CreateVariantRequest'
          nullable: true
    RecipeAiFlags:
      type: object
      required: [generativeAdjustmentsEnabled]
      properties:
        generativeAdjustmentsEnabled:
          type: boolean
    ApplyProfileRequest:
      type: object
      properties:
        dairyMode:
          type: string
          description: none | omit | substitute_oat
        omitTokens:
          type: array
          items:
            type: string
        useGenerative:
          type: boolean
          description: Opt-in; server rejects when feature flag is off
    ApplyProfileResponse:
      type: object
      required: [adjustmentId, appliedProfile, summary, adjustedIngredients, adjustedSteps]
      properties:
        adjustmentId:
          type: string
          format: uuid
        appliedProfile:
          type: object
          additionalProperties: true
        summary:
          type: string
        adjustedIngredients:
          type: array
          items:
            $ref: '#/components/schemas/IngredientLineDto'
        adjustedSteps:
          type: array
          items:
            $ref: '#/components/schemas/RecipeStepDto'
    VariantDetailResponse:
      type: object
      required: [id, dishId, title, canonical, ingredients, steps]
      properties:
        id:
          type: string
          format: uuid
        dishId:
          type: string
          format: uuid
        title:
          type: string
        yields:
          type: string
          nullable: true
        prepTimeMin:
          type: integer
          nullable: true
        cookTimeMin:
          type: integer
          nullable: true
        canonical:
          type: boolean
        sourceId:
          type: string
          format: uuid
          nullable: true
        ownerUserId:
          type: string
          format: uuid
          nullable: true
        ingredients:
          type: array
          items:
            $ref: '#/components/schemas/IngredientLineDto'
        steps:
          type: array
          items:
            $ref: '#/components/schemas/RecipeStepDto'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ProblemDetails:
      type: object
      required: [title, status]
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string